application properties

+++ src/main/java/salao/online/application/dtos/dtosDeCliente/AtualizarClienteDTO.java
@@ -17,7 +17,7 @@ public class AtualizarClienteDTO{
     private String nome;
     private String sobrenome;
     private String nomeSocial;
-    private int idade;
+    private short idade;
     private String email;
     private String telefone;
     private String senha;
